+++ gossip/src/lib.rs
@@ -34,6 +34,8 @@ pub mod restart_crds_values;
 mod verifying_key_cache;
 pub mod weighted_shuffle;
 
+pub use solana_net_utils::PinnedXdpSender as XdpSender;
+
 #[macro_use]
 extern crate log;