+++ gossip/tests/gossip.rs
@@ -46,6 +46,7 @@ fn test_node(exit: Arc<AtomicBool>) -> (Arc<ClusterInfo>, GossipService, UdpSock
         None,
         test_node.sockets.gossip,
         None,
+        None,
         true, // should_check_duplicate_instance
         None,
         exit,
@@ -73,6 +74,7 @@ fn test_node_with_bank(
         &cluster_info,
         Some(epoch_specs),
         test_node.sockets.gossip,
+        None,
         None,
         true, // should_check_duplicate_instance
         None,